refactor(suite-native): move RoundedIcon into Icon directory
