+++ suite-native/atoms/src/index.ts
@@ -60,7 +60,7 @@ export * from './Pictogram/Pictogram';
 export * from './TitleHeader/PictogramTitleHeader';
 export * from './TitleHeader/TitleHeader';
 export * from './TitleHeader/CenteredTitleHeader';
-export * from './RoundedIcon';
+export * from './Icon/RoundedIcon';
 export * from './CompoundRoundedIcon';
 export * from './TrezorSuiteHeader';
 export * from './Skeleton/BoxSkeleton';
